refactor: split schema/ and model/, and retire the import cycle with them
Phase A of the 3.7.0 refactoring (spec/007_architecture.md § schema/ versus model/).
schema/ is what crosses the network; model/ is what the plugin owns and persists locally.

The split is worth doing because the codebase already drew the line and then ignored it.
schema/base.py gives Serializable an as_dict/as_json for request bodies and SkipDataClass
a from_dict whose docstring says response parsing should inherit it. Two things did not
fit that story: the provider classes, which the user configures and which persist under
mapflow_data_providers, and ProcessingHistory, which reads and writes
processing_history_{project_id} in QgsSettings. Neither ever touches the wire. Both are
now in model/, and that is the whole of model/ - it is meant to stay small.

The payoff is that the import cycle is gone as a category rather than patched. The rule
that model/ may import schema/ and schema/ may never import model/ makes the old cycle
unrepresentable: schema/processing.py reaching into the provider package for SourceType
was exactly a violation of that direction. SourceType, CRS and BasicAuth are wire
vocabulary - they go into request bodies - so they belong in schema/provider_types.py, and
schema/processing.py now imports a stdlib-only leaf instead of a package.

Worth recording precisely, because the conftest comments had it wrong and cost someone a
wrong mental model: the cycle was never a four-hop chain through layer_utils and dialogs.
entity/provider/provider.py imported nothing from the plugin at all. The cycle came from
package initialisation - importing entity.provider.provider first runs the package
__init__, which imports basemap_provider, which imports back into the half-built
schema.processing. Both test tiers imported the tree twice and swallowed the first
ImportError to get past it. Those retry loops are deleted, and both tiers pass without
them, which is the proof the cycle is structurally gone rather than retried around.

default.py and basemap_provider.py now import the three wire types straight from
schema.provider_types rather than picking them up from model/provider/provider.py, which
re-exports them for its own use. Same objects either way; the point is that a reader of
those files sees which side of the boundary the type lives on.

## Manual test

Regression surface - this moves types that every feature touches, so the risk is an import
that only fails on a path the tests do not reach. Worth one pass through the plugin:
- providers: open settings, add an XYZ and a TMS provider with credentials, save, reopen
  QGIS and confirm both survive with credentials intact - ProvidersList is the settings
  round-trip most affected
- start a processing with a Mapflow provider, with My Imagery, and with a search result,
  since to_processing_params builds the request body from the moved SourceType/CRS
- imagery search: run one, confirm footprints load and preview works
- open a project with existing processings and confirm the table populates - that reads
  ProcessingHistory back from settings written by an earlier build

New behaviour: none. No runtime behaviour changes; every change is a move or an import
rewrite.
