+++ mapflow/model/provider/__init__.py
@@ -2,4 +2,5 @@ from .basemap_provider import XYZProvider, TMSProvider, QuadkeyProvider
 from .collection import ProvidersList
 from .default import DefaultProvider, ImagerySearchProvider, MyImageryProvider
 from .factory import create_provider
-from .provider import UsersProvider, CRS, SourceType, BasicAuth, ProviderInterface
+from .provider import UsersProvider, ProviderInterface
+from ...schema.provider_types import CRS, SourceType, BasicAuth
