+++ tests/qgis/conftest.py
@@ -4,8 +4,6 @@ Tests in this directory need a real PyQGIS runtime — they import plugin
 modules that touch qgis.core / qgis.gui at module load time. Run inside
 the qgis/qgis:release-3_28 Docker image (see Dockerfile.tests + Makefile).
 """
-import importlib
-
 import pytest
 from unittest.mock import MagicMock
 
@@ -19,17 +17,6 @@ def pytest_configure(config):
     """
     from qgis.testing import start_app
     start_app()
-
-    # Pre-warm the mapflow module tree to survive the circular import on first load.
-    # The chain mapflow.schema.processing -> entity.provider -> functional.layer_utils
-    # -> dialogs -> mapflow.schema creates a circular dependency that fails on the
-    # first attempt but succeeds on retry because partial modules are cached.
-    for _ in range(2):
-        try:
-            importlib.import_module("mapflow.schema.processing")
-            break
-        except ImportError:
-            pass
 
 
 @pytest.fixture()
